Turn warning into errors
Closes #19

+++ src/vmod_digest.c
@@ -246,7 +246,7 @@ vmod_hmac_generic(VRT_CTX, hashid hash, const char *key, const char *msg)
 	unsigned char mac[blocksize];
 	unsigned char *hexenc;
 	unsigned char *hexptr;
-	int j;
+	unsigned j;
 	MHASH td;
 
 	assert(msg);
@@ -280,7 +280,7 @@ vmod_hmac_generic(VRT_CTX, hashid hash, const char *key, const char *msg)
 	for (j = 0; j < blocksize; j++) {
 		sprintf((char*)hexptr,"%.2x", mac[j]);
 		hexptr+=2;
-		assert((hexptr-hexenc)<(2*blocksize + 3));
+		assert((size_t)(hexptr - hexenc) < (2 * blocksize + 3));
 	}
 	*hexptr = '\0';
 	return (const char *)hexenc;
@@ -337,7 +337,7 @@ vmod_hash_generic(VRT_CTX, hashid hash, const char *msg)
 {
 	MHASH td;
 	unsigned char h[mhash_get_block_size(hash)];
-	int i;
+	unsigned i;
 	char *p;
 	char *ptmp;
 
@@ -348,7 +348,7 @@ vmod_hash_generic(VRT_CTX, hashid hash, const char *msg)
 	p = WS_Alloc(ctx->ws,mhash_get_block_size(hash)*2 + 1);
 	AN(p);
 	ptmp = p;
-	for (i = 0; i<mhash_get_block_size(hash);i++) {
+	for (i = 0; i < mhash_get_block_size(hash); i++) {
 		sprintf(ptmp,"%.2x",h[i]);
 		ptmp+=2;
 	}
